v1.1/requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.4.0
psutil>=5.9.0